requirements.txt
langchain==0.1.14
langchain-google-genai==0.0.8
langchain-community==0.0.28
google-generativeai==0.4.1
chromadb==0.4.24
huggingface-hub==0.22.2
sentence-transformers==2.6.1
streamlit==1.33.0
python-dotenv==1.0.1